Change the toolbar icon when connecting
This change adds a member to track whether there is a device connecting.
Each time the model changes we iterate to see if there are any devices
connecting and set the member variable accordingly. When the boolean
connecting variable changes a signal "status-connecting" is emitted with
the value of the connecting member.

https://bugzilla.gnome.org/show_bug.cgi?id=597811