Fix crash caused by invalid cast
I badly merged some branches before an earlier patch which resulted in a
GtkButton being sent as the data to a callback which I was then casting to
a MoblinPanel.

This patch fixes that and aims to alleviate this in future by consistently
passing self (a pointer to MoblinPanel) as the callback data.

https://bugzilla.gnome.org/show_bug.cgi?id=597910